requirements.txt
sqlalchemy
flasgger
psycopg2
flask-cors
flask_sqlalchemy
